Updated auth middleware structure.
A single auth middleware that was being applied to all routes,
has now been broken up into two seperate middlewares,
one for initial users and the other one for normal users.

From now on,auth middleware will be applied to individual routes as per requirement.
Another middleware know as enableOptionalAuth has been introduced, which has to be applied before
the auth middlware in the chain. This middleware allows the request to bypass authorization if access_token is not present in the
authorization header.